Don't delete AuthPolicy/TokenRateLimitPolicy with opendatahub.io/managed=false when deleting the MaasModel object

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED

+++ maas-controller/pkg/controller/maas/maasmodel_controller.go
@@ -199,6 +199,11 @@ func (r *MaaSModelReconciler) deleteGeneratedPoliciesByLabel(ctx context.Context
 
 	for i := range policyList.Items {
 		p := &policyList.Items[i]
+		if isManaged(p) {
+			log.Info(fmt.Sprintf("Generated %s opted out, skipping deletion", kind),
+				"name", p.GetName(), "namespace", p.GetNamespace(), "model", modelName)
+			continue
+		}
 		log.Info(fmt.Sprintf("Deleting generated %s on MaaSModel deletion", kind),
 			"name", p.GetName(), "namespace", p.GetNamespace(), "model", modelName)
 		if err := r.Delete(ctx, p); err != nil && !apierrors.IsNotFound(err) {
